Extract Mermaid stuff to a separate class
